Renamed openquake.commonlib.calculators -> openquake.calculators
