+++ celeryconfig.py
@@ -79,13 +79,13 @@ CELERY_ACCEPT_CONTENT = ['pickle', 'json']
 CELERY_IMPORTS = get_core_modules(engine) + [
     "openquake.engine.calculators.hazard.general",
     "openquake.engine.tests.utils.tasks"] + [
-    "openquake.commonlib.calculators.classical",
-    "openquake.commonlib.calculators.classical_risk",
-    "openquake.commonlib.calculators.classical_damage",
-    "openquake.commonlib.calculators.event_based",
-    "openquake.commonlib.calculators.event_based_risk",
-    "openquake.commonlib.calculators.scenario_risk",
-    "openquake.commonlib.calculators.scenario_damage",
+    "openquake.calculators.classical",
+    "openquake.calculators.classical_risk",
+    "openquake.calculators.classical_damage",
+    "openquake.calculators.event_based",
+    "openquake.calculators.event_based_risk",
+    "openquake.calculators.scenario_risk",
+    "openquake.calculators.scenario_damage",
     ]
 
 os.environ["DJANGO_SETTINGS_MODULE"] = "openquake.engine.settings"
