+++ openquake/engine/calculators/hazard/disaggregation/core.py
@@ -28,7 +28,7 @@ from openquake.hazardlib.imt import from_string
 from openquake.hazardlib.site import SiteCollection
 
 from openquake.baselib.general import groupby
-from openquake.commonlib.calculators.calc import gen_ruptures_for_site
+from openquake.calculators.calc import gen_ruptures_for_site
 
 from openquake.engine import logs
 from openquake.engine.db import models
